Implement methods delegating to fold methods
